moves class to reflection package instead of reflection
