+++ artemis-core/src/main/java/org/jnosql/artemis/reflection/JavaCompiledStream.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.reflection.compiler;
+package org.jnosql.artemis.reflection;
 
 
 import java.io.ByteArrayInputStream;
