+++ artemis-core/src/main/java/org/jnosql/artemis/reflection/TemplateReader.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.reflection.compiler;
+package org.jnosql.artemis.reflection;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
