+++ artemis-core/src/test/java/org/jnosql/artemis/reflection/JavaCompilerBeanPropertyReaderFactory.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.jnosql.artemis.reflection.compiler;
+package org.jnosql.artemis.reflection;
 
 class JavaCompilerBeanPropertyReaderFactory {
