--- artemis-core/src/test/java/org/jnosql/artemis/reflection/compiler/Foo.java
@@ -1,28 +1,0 @@
-/*
- *  Copyright (c) 2017 Otávio Santana and others
- *   All rights reserved. This program and the accompanying materials
- *   are made available under the terms of the Eclipse Public License v1.0
- *   and Apache License v2.0 which accompanies this distribution.
- *   The Eclipse Public License is available at http://www.eclipse.org/legal/epl-v10.html
- *   and the Apache License v2.0 is available at http://www.opensource.org/licenses/apache2.0.php.
- *
- *   You may elect to redistribute this code under either of these licenses.
- *
- *   Contributors:
- *
- *   Otavio Santana
- */
-package org.jnosql.artemis.reflection.compiler;
-
- public class Foo {
-
-     private String bar;
-
-     public String getBar() {
-         return bar;
-     }
-
-     public void setBar(String bar) {
-         this.bar = bar;
-     }
- }
